requirements.txt
pyspark>=3.5.0
requests>=2.31.0
python-dotenv>=1.0.0
pandas>=2.1.0
pytest>=7.4.0
pyarrow>=14.0.0
sqlalchemy>=2.0.0
